forward initial value as 64bit value

+++ heob.h
@@ -443,7 +443,7 @@ typedef LONG NTAPI func_NtQueryInformationThread(
 // }}}
 // common functions {{{
 
-char *num2hexstr( char *str,uintptr_t arg,int count );
+char *num2hexstr( char *str,UINT64 arg,int count );
 char *num2str( char *start,uintptr_t arg,int minus );
 char *mstrrchr( const char *s,char c );
 int isWrongArch( HANDLE process );
